Fixed bug on unreliable connection when avatars are removed from the scene
Whenever object connection fails script waits 10 seconds for object to
reconnect. Waiting is done using setTimeout. If the object is connected
before time is out, the object is not to be deleted. However, timeout was
not properly cleared which resulted in removal of avatars on laggy
connections, despite the fact that they reconnect within 10 seconds.